Remove omrmmap.c and omrmmap.h for zos390 directory
