+++ port/unix/omrmmap.c
@@ -50,6 +50,10 @@
 #include "protect_helpers.h"
 #include "omrportpriv.h"
 
+#if defined(J9ZOS390)
+#include "omrvmem.h"
+#endif /* defined(J9ZOS390) */
+
 #if defined(AIXPPC)
 #include <sys/shm.h>
 #include <sys/vminfo.h>
@@ -79,6 +83,81 @@
 J9MmapHandle *
 omrmmap_map_file(struct OMRPortLibrary *portLibrary, intptr_t file, uint64_t offset, uintptr_t size, const char *mappingName, uint32_t flags, uint32_t categoryCode)
 {
+#if defined(J9ZOS390)
+	/* default implementation will allocate memory and read the file into it */
+	uintptr_t numBytesRead;
+	intptr_t rc;
+	void *mappedMemory;
+	void *allocPointer;
+	J9MmapHandle *returnVal;
+
+	Trc_PRT_mmap_map_file_default_entered(file, size);
+
+	if (file == -1) {
+		Trc_PRT_mmap_map_file_default_badfile();
+		return NULL;
+	}
+
+	if (0 == size) {
+		struct stat buf;
+
+		memset(&buf, 0, sizeof(struct stat));
+		if (-1 == fstat(file - FD_BIAS, &buf)) {
+			Trc_PRT_mmap_map_file_unix_filestatfailed();
+			portLibrary->error_set_last_error(portLibrary, errno, OMRPORT_ERROR_MMAP_MAP_FILE_STATFAILED);
+			return NULL;
+		}
+		size = buf.st_size;
+	}
+
+	if (0 != portLibrary->file_seek(portLibrary, file, offset, EsSeekSet)) {
+		Trc_PRT_mmap_map_seek_failed(offset);
+		return NULL;
+	}
+
+	/* ensure that allocated memory is 8 byte aligned, just in case it matters */
+	allocPointer = portLibrary->mem_allocate_memory(portLibrary, size + 8, OMR_GET_CALLSITE(), categoryCode);
+	Trc_PRT_mmap_map_file_default_allocPointer(allocPointer, size + 8);
+
+	if (allocPointer == NULL) {
+		Trc_PRT_mmap_map_file_default_badallocPointer();
+		portLibrary->file_close(portLibrary, file);
+		return NULL;
+	}
+
+	if ((uintptr_t)allocPointer % 8) {
+		mappedMemory = (void *)((uintptr_t)allocPointer + 8 - ((uintptr_t)allocPointer % 8));
+	} else {
+		mappedMemory = allocPointer;
+	}
+	Trc_PRT_mmap_map_file_default_mappedMemory(mappedMemory);
+
+	numBytesRead = 0;
+	rc = 0;
+	while (size - numBytesRead > 0) {
+		rc = portLibrary->file_read(portLibrary, file, (void *)(((uintptr_t)mappedMemory) + numBytesRead), size - numBytesRead);
+		if (rc == -1) {
+			/* failed to completely read the file */
+			Trc_PRT_mmap_map_file_default_badread();
+			portLibrary->mem_free_memory(portLibrary, allocPointer);
+			return NULL;
+		}
+		numBytesRead += rc;
+		Trc_PRT_mmap_map_file_default_readingFile(numBytesRead);
+	}
+
+	if (!(returnVal = (J9MmapHandle *)portLibrary->mem_allocate_memory(portLibrary, sizeof(J9MmapHandle), OMR_GET_CALLSITE(), categoryCode))) {
+		Trc_PRT_mmap_map_file_cannotallocatehandle();
+		portLibrary->mem_free_memory(portLibrary, allocPointer);
+		return NULL;
+	}
+	returnVal->pointer = mappedMemory;
+	returnVal->allocPointer = allocPointer;
+	returnVal->size = size;
+
+	Trc_PRT_mmap_map_file_default_exit();
+	return returnVal;
+#else /* defined(J9ZOS390) */
 	int mmapProt = 0;
 	int mmapFlags = 0;
 	void *pointer = NULL;
@@ -170,6 +249,7 @@ omrmmap_map_file(struct OMRPortLibrary *portLibrary, intptr_t file, uint64_t off
 	/* Completed, return */
 	Trc_PRT_mmap_map_file_unix_exiting(pointer, returnVal);
 	return returnVal;
+#endif /* defined(J9ZOS390) */
 }
 /**
  * UnMap previously mapped memory.
@@ -181,6 +261,12 @@ omrmmap_map_file(struct OMRPortLibrary *portLibrary, intptr_t file, uint64_t off
 void
 omrmmap_unmap_file(struct OMRPortLibrary *portLibrary, J9MmapHandle *handle)
 {
+#if defined(J9ZOS390)
+	if (handle != NULL) {
+		portLibrary->mem_free_memory(portLibrary, handle->allocPointer);
+		portLibrary->mem_free_memory(portLibrary, handle);
+	}
+#else /* defined(J9ZOS390) */
 	intptr_t rc = 0;
 
 	Trc_PRT_mmap_unmap_file_unix_entering(handle);
@@ -193,6 +279,7 @@ omrmmap_unmap_file(struct OMRPortLibrary *portLibrary, J9MmapHandle *handle)
 	}
 
 	Trc_PRT_mmap_unmap_file_unix_exiting(rc);
+#endif /* defined(J9ZOS390) */
 }
 /**
  * Synchronise updates to memory mapped file region with file on disk.  The call may wait for the file write
@@ -216,6 +303,10 @@ omrmmap_unmap_file(struct OMRPortLibrary *portLibrary, J9MmapHandle *handle)
 intptr_t
 omrmmap_msync(struct OMRPortLibrary *portLibrary, void *start, uintptr_t length, uint32_t flags)
 {
+#if defined(J9ZOS390)
+	Trc_PRT_mmap_msync_default_entered(start, length, flags);
+	return -1;
+#else /* defined(J9ZOS390) */
 	int msyncFlags = 0;
 	int rc = 0;
 
@@ -248,6 +339,7 @@ omrmmap_msync(struct OMRPortLibrary *portLibrary, void *start, uintptr_t length,
 
 	Trc_PRT_mmap_msync_unix_exiting();
 	return 0;
+#endif /* defined(J9ZOS390) */
 }
 /**
  * PortLibrary shutdown.
@@ -308,8 +400,10 @@ omrmmap_capabilities(struct OMRPortLibrary *portLibrary)
 {
 	return (OMRPORT_MMAP_CAPABILITY_COPYONWRITE
 			| OMRPORT_MMAP_CAPABILITY_READ
+#if !defined(J9ZOS390)
 			| OMRPORT_MMAP_CAPABILITY_PROTECT
-			/* If JSE platforms include WRITE and MSYNC - ZOS included, but currently has own omrmmap.c */
+#endif /* defined(J9ZOS390) */
+			/* If JSE platforms include WRITE and MSYNC  */
 #if ((defined(LINUX) && defined(J9X86)) \
   || (defined(LINUXPPC)) \
   || (defined(LINUX) && defined(S390)) \
@@ -328,14 +422,31 @@ omrmmap_capabilities(struct OMRPortLibrary *portLibrary)
 intptr_t
 omrmmap_protect(struct OMRPortLibrary *portLibrary, void *address, uintptr_t length, uintptr_t flags)
 {
+#if defined(J9ZOS390)
+	return OMRPORT_PAGE_PROTECT_NOT_SUPPORTED;
+#else /* defined(J9ZOS390) */
 	return protect_memory(portLibrary, address, length, flags);
+#endif /* defined(J9ZOS390) */
 }
 
 uintptr_t
 omrmmap_get_region_granularity(struct OMRPortLibrary *portLibrary, void *address)
 {
+#if defined(J9ZOS390)
+	return 0;
+#else /* defined(J9ZOS390) */
 	return protect_region_granularity(portLibrary, address);
+#endif /* defined(J9ZOS390) */
 }
+
+#if defined(J9ZOS390)
+#pragma linkage (PGSERRM,OS)
+#pragma map(Pgser_Release,"PGSERRM")
+#if defined(OMR_ENV_DATA64)
+#pragma linkage(omrdiscard_data,OS_NOSTACK)
+int omrdiscard_data(void *address, int numFrames);
+#endif /*OMR_ENV_DATA64 */
+#endif /* defined(J9ZOS390) */
 
 /**
  * Disclaim entire pages only: don't release partial pages at the start and end.
@@ -353,9 +464,7 @@ omrmmap_dont_need(struct OMRPortLibrary *portLibrary, const void *startAddress, 
 		uintptr_t roundedStart = ROUND_UP_TO_POWEROF2((uintptr_t) startAddress, pageSize);
 		size_t roundedLength = ROUND_DOWN_TO_POWEROF2(endAddress - roundedStart, pageSize);
 		if (roundedLength >= pageSize) {
-
 			Trc_PRT_mmap_dont_need_oscall(roundedStart, roundedLength);
-
 #if defined(LINUX) || defined(OSX)
 			if (-1 == madvise((void *)roundedStart, roundedLength, MADV_DONTNEED)) {
 				Trc_PRT_mmap_dont_need_madvise_failed((void *)roundedStart, roundedLength, errno);
@@ -365,7 +474,17 @@ omrmmap_dont_need(struct OMRPortLibrary *portLibrary, const void *startAddress, 
 			if (-1 == disclaim64((void *)roundedStart, roundedLength, DISCLAIM_ZEROMEM)) {
 				Trc_PRT_mmap_dont_need_disclaim64_failed((void *)roundedStart, roundedLength, errno);
 			}
-#endif /* defined(AIXPPC) */
+#elif defined(J9ZOS390)
+#if defined(OMR_ENV_DATA64)
+			if (omrdiscard_data((void *)roundedStart, roundedLength >> ZOS_REAL_FRAME_SIZE_SHIFT) != 0) {
+				Trc_PRT_mmap_dont_need_j9discard_data_failed((void *)roundedStart, roundedLength);
+			}
+#else
+			if (Pgser_Release((void *)roundedStart, roundedLength) != 0) {
+				Trc_PRT_mmap_dont_need_Pgser_Release_failed((void *)roundedStart, roundedLength);
+			}
+#endif /* defined(OMR_ENV_DATA64) */
+#endif /* defined(J9ZOS390) */
 		}
 	}
 }
