--- port/zos390/omrmmap.h
@@ -1,36 +1,0 @@
-/*******************************************************************************
- * Copyright IBM Corp. and others 1991
- *
- * This program and the accompanying materials are made available under
- * the terms of the Eclipse Public License 2.0 which accompanies this
- * distribution and is available at https://www.eclipse.org/legal/epl-2.0/
- * or the Apache License, Version 2.0 which accompanies this distribution and
- * is available at https://www.apache.org/licenses/LICENSE-2.0.
- *
- * This Source Code may also be made available under the following
- * Secondary Licenses when the conditions for such availability set
- * forth in the Eclipse Public License, v. 2.0 are satisfied: GNU
- * General Public License, version 2 with the GNU Classpath
- * Exception [1] and GNU General Public License, version 2 with the
- * OpenJDK Assembly Exception [2].
- *
- * [1] https://www.gnu.org/software/classpath/license.html
- * [2] https://openjdk.org/legal/assembly-exception.html
- *
- * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0 OR GPL-2.0 WITH Classpath-exception-2.0 OR LicenseRef-GPL-2.0 WITH Assembly-exception
- *******************************************************************************/
-
-#ifndef omrmmap_h
-#define omrmmap_h
-
-void omrmmap_unmap_file(struct OMRPortLibrary *portLibrary, J9MmapHandle *handle);
-J9MmapHandle *omrmmap_map_file(struct OMRPortLibrary *portLibrary, intptr_t file, uint64_t offset, uintptr_t size, const char *mappingName, uint32_t flags);
-int32_t omrmmap_startup(struct OMRPortLibrary *portLibrary);
-int32_t omrmmap_capabilities(struct OMRPortLibrary *portLibrary);
-void omrmmap_shutdown(struct OMRPortLibrary *portLibrary);
-intptr_t omrmmap_msync(struct OMRPortLibrary *portLibrary, void *start, uintptr_t length, uint32_t flags);
-
-
-#endif     /* omrmmap_h */
-
-
